chore: extract dayz from valve
