+++ lib/games.js
@@ -654,7 +654,7 @@ export const games = {
     options: {
       port: 2302,
       port_query_offset: 24714,
-      protocol: 'valve'
+      protocol: 'dayz'
     }
   },
   dayzmod: {
