+++ protocols/index.js
@@ -50,11 +50,12 @@ import ventrilo from './ventrilo.js'
 import warsow from './warsow.js'
 import beammpmaster from './beammpmaster.js'
 import beammp from './beammp.js'
+import dayz from './dayz.js'
 
 export {
   armagetron, ase, asa, assettocorsa, battlefield, buildandshoot, cs2d, discord, doom3, eco, epic, ffow, fivem, gamespy1,
   gamespy2, gamespy3, geneshift, goldsrc, hexen2, jc2mp, kspdmp, mafia2mp, mafia2online, minecraft,
   minecraftbedrock, minecraftvanilla, mumble, mumbleping, nadeo, openttd, quake1, quake2, quake3, rfactor, samp,
   savage2, starmade, starsiege, teamspeak2, teamspeak3, terraria, tribes1, tribes1master, unreal2, ut3, valve,
-  vcmp, ventrilo, warsow, eldewrito, beammpmaster, beammp
+  vcmp, ventrilo, warsow, eldewrito, beammpmaster, beammp, dayz
 }
